Customer Retention in Telecom/TelecomRetainGPT/backend/requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
pypdf
pdfminer.six
python-docx
pdf2image
Pillow
pytesseract
python-dotenv
scikit-learn
numpy
